[fix] moved wait method
